Moved collections to independent JSON files
