+++ src/assets/Collections/Dishes.json
@@ -1,0 +1,55 @@
+[
+        [
+                { "Name": "Creamy Carrot Soup", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Fish Stew", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Fish Tacos", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Grilled Fish", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Grilled Meat", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Grilled Mushroom", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Grilled Oyster", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Hearty Vegetable Soup", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Hot Hounds", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Macaron", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Palian Onion Soup", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Sashimi", "status": "No", "additionalDetails": "placeholder"}
+        ],
+        [
+                { "Name": "Bacon-Stuffed Mushrooms", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Blueberry Pie", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Chapaa Asada Tacos", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Crab Gumbo", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Crab Pot Pie", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Cream of Mushroom Soup", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Cream of Tomato Soup", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Fisherman's Brew", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Fried Catfish Dinner", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Loaded Potato Soup", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Ramen", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Sernuk Noodle Stew", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Shepp's Pie", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Steak Dinner", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Trout Dinner", "status": "No", "additionalDetails": "placeholder"}
+        ],
+        [
+                { "Name": "Akwinduu Chapaa", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Apple Pie", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Bouillabaisse", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Chapaa Onigiri", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Chili Oil Dumplings", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Meaty Stir Fry", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Phoenixfire Relleno", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Spicy Stir Fry", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Sushi", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Veggie Fried Rice", "status": "No", "additionalDetails": "placeholder"}
+        ],
+        [
+                { "Name": "Celebration Cake", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Chapaa Masala", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Poke Bowl", "status": "No", "additionalDetails": "placeholder"},
+                { "Name": "Spicy Rice Cakes", "status": "No", "additionalDetails": "placeholder"}
+        ],
+                { "Name": "Petit Fives", "status": "No", "additionalDetails": "placeholder" },
+                { "Name": "Oysters Akwinduu", "status": "No", "additionalDetails": "placeholder" },
+                { "Name": "Muujin Bahari", "status": "No", "additionalDetails": "placeholder" },
+                { "Name": "Congee", "status": "No", "additionalDetails": "placeholder" }
+]
