--- src/assets/collections.js
@@ -1,1148 +1,0 @@
-const Villagers =       [
-                            {Name: 'Ashura', dailyGift: false, romance: false, weeklyRomance: false, weeklyWants: [
-                                                                                                        {Name: "Grilled Fish", Gifted: false},
-                                                                                                        {Name: "Kilima Catfish", Gifted: false},
-                                                                                                        {Name: "Crab Pot Pie", Gifted: false},
-                                                                                                        {Name: "Fisherman's Brew", Gifted: false}
-                                                                                                    ]},
-                            {Name: 'Auni', dailyGift: false, romance: false, weeklyRomance: false, weeklyWants:    [
-                                                                                                        {Name: "Unopened Oyster", Gifted: false}, 
-                                                                                                        {Name: "Garden Mantis", Gifted: false}, 
-                                                                                                        {Name: "Spotted Mantis", Gifted: false}, 
-                                                                                                        {Name: "Apple Pie", Gifted: false}
-                                                                                                    ]},
-                            {Name: 'Badruu', dailyGift: false, romance: false, weeklyRomance: false, weeklyWants:  [
-                                                                                                        {Name: "Carrot Seed", Gifted: false}, 
-                                                                                                        {Name: "Wheat", Gifted: false}, 
-                                                                                                        {Name: "Blueberry Bush Seed", Gifted: false}, 
-                                                                                                        {Name: "Meaty Stir Fry", Gifted: false}
-                                                                                                    ]},
-                            {Name: 'Caleri', dailyGift: false, romance: false, weeklyRomance: false, weeklyWants:  [
-                                                                                                        {Name: "Coral", Gifted: false}, 
-                                                                                                        {Name: "Pickled Carrots", Gifted: false}, 
-                                                                                                        {Name: "Silk Thread", Gifted: false}, 
-                                                                                                        {Name: "Dari Cloves", Gifted: false}
-                                                                                                    ]},
-                            {Name: 'Chayne', dailyGift: false, romance: false, weeklyRomance: false, weeklyWants:  [
-                                                                                                        {Name: "Potato Seed", Gifted: false}, 
-                                                                                                        {Name: "Juniper Seed", Gifted: false}, 
-                                                                                                        {Name: "Blueberries", Gifted: false}, 
-                                                                                                        {Name: "Palian Onion Soup", Gifted: false}
-                                                                                                    ]},
-                            {Name: 'Delaila', dailyGift: false, romance: false, weeklyRomance: false, weeklyWants: [
-                                                                                                        {Name: "Clay", Gifted: false}, 
-                                                                                                        {Name: "Wheat Seed", Gifted: false}, 
-                                                                                                        {Name: "Shepp's Pie", Gifted: false}, 
-                                                                                                        {Name: "Gold Bar", Gifted: false}
-                                                                                                    ]},
-                            {Name: 'Einar', dailyGift: false, romance: true, weeklyRomance: false, weeklyWants:   [
-                                                                                                        {Name: "Worm", Gifted: false}, 
-                                                                                                        {Name: "Ribbontail Ray", Gifted: false}, 
-                                                                                                        {Name: "Mottled Gobi", Gifted: false}, 
-                                                                                                        {Name: "Radiant Sunfish", Gifted: false}
-                                                                                                    ]},
-                            {Name: 'Elouisa', dailyGift: false, romance: false, weeklyRomance: false, weeklyWants: [
-                                                                                                        {Name: "Duskray", Gifted: false}, 
-                                                                                                        {Name: "Inky Dragonfly", Gifted: false}, 
-                                                                                                        {Name: "Blobfish", Gifted: false}, 
-                                                                                                        {Name: "Giant Goldfish", Gifted: false}
-                                                                                                    ]},
-                            {Name: 'Eshe', dailyGift: false, romance: false, weeklyRomance: false, weeklyWants:    [
-                                                                                                        {Name: "Fur", Gifted: false}, 
-                                                                                                        {Name: "Bluefin Tuna", Gifted: false}, 
-                                                                                                        {Name: "Macaron", Gifted: false}, 
-                                                                                                        {Name: "Palian Onion Soup", Gifted: false}
-                                                                                                    ]},
-                            {Name: 'Hassian', dailyGift: false, romance: true, weeklyRomance: false, weeklyWants: [
-                                                                                                        {Name: "Fur", Gifted: false}, 
-                                                                                                        {Name: "Chapaa Asada Tacos", Gifted: false}, 
-                                                                                                        {Name: "Proudhorned Sernuk Antlers", Gifted: false}, 
-                                                                                                        {Name: "Slowdown Arrow", Gifted: false}
-                                                                                                    ]},
-                            {Name: 'Hekla', dailyGift: false, romance: false, weeklyRomance: false, weeklyWants:   [
-                                                                                                        {Name: "Grilled Meat", Gifted: false}, 
-                                                                                                        {Name: "Bahari Bream", Gifted: false}, 
-                                                                                                        {Name: "Black Sea Bass", Gifted: false}, 
-                                                                                                        {Name: "Green Pearl", Gifted: false}
-                                                                                                    ]},
-                            {Name: 'Hodari', dailyGift: false, romance: true, weeklyRomance: false, weeklyWants:  [
-                                                                                                        {Name: "Spotted Chapaa Tail", Gifted: false}, 
-                                                                                                        {Name: "Copper Bar", Gifted: false}, 
-                                                                                                        {Name: "Stonefish", Gifted: false}, 
-                                                                                                        {Name: "Spitfire Cicada", Gifted: false}
-                                                                                                    ]},
-                            {Name: 'Jel', dailyGift: false, romance: true, weeklyRomance: false, weeklyWants:     [
-                                                                                                        {Name: "Coral", Gifted: false}, 
-                                                                                                        {Name: "Freshwater Eel", Gifted: false}, 
-                                                                                                        {Name: "Petit Fives", Gifted: false}, 
-                                                                                                        {Name: "Bluebristle Muujin Mane", Gifted: false}
-                                                                                                    ]},
-                            {Name: 'Jina', dailyGift: false, romance: true, weeklyRomance: false, weeklyWants:    [
-                                                                                                        {Name: "Wagon Wheel", Gifted: false}, 
-                                                                                                        {Name: "Brightshroom", Gifted: false}, 
-                                                                                                        {Name: "Radiant Sunfish", Gifted: false}, 
-                                                                                                        {Name: "Rainbow-Tipped Butterfly", Gifted: false}
-                                                                                                    ]},
-                            {Name: 'Kenli', dailyGift: false, romance: false, weeklyRomance: false, weeklyWants:   [
-                                                                                                        {Name: "Spotted Chapaa Tail", Gifted: false}, 
-                                                                                                        {Name: "Ramen", Gifted: false}, 
-                                                                                                        {Name: "Trout Dinner", Gifted: false}, 
-                                                                                                        {Name: "Meaty Stir Fry", Gifted: false}
-                                                                                                    ]},
-                            {Name: 'Kenyatta', dailyGift: false, romance: true, weeklyRomance: false, weeklyWants:[
-                                                                                                        {Name: "Crystal Lake Lotus", Gifted: false}, 
-                                                                                                        {Name: "Cutthroat Trout", Gifted: false}, 
-                                                                                                        {Name: "Firebreathing Dragonfly", Gifted: false}, 
-                                                                                                        {Name: "Bluebristle Muujin Mane", Gifted: false}
-                                                                                                    ]},
-                            {Name: 'Nai\'O', dailyGift: false, romance: true, weeklyRomance: false, weeklyWants:  [
-                                                                                                        {Name: "Wagon Wheel", Gifted: false}, 
-                                                                                                        {Name: "Kilima Catfish", Gifted: false}, 
-                                                                                                        {Name: "Bahari Pike", Gifted: false}, 
-                                                                                                        {Name: "Steak Dinner", Gifted: false}
-                                                                                                    ]},
-                            {Name: 'Najuma', dailyGift: false, romance: false, weeklyRomance: false, weeklyWants:  [
-                                                                                                        {Name: "Tomato Plant Seed", Gifted: false}, 
-                                                                                                        {Name: "Copper Bar", Gifted: false}, 
-                                                                                                        {Name: "Apple", Gifted: false}, 
-                                                                                                        {Name: "Blueberry Pie", Gifted: false}
-                                                                                                    ]},
-                            {Name: 'Reth', dailyGift: false, romance: true, weeklyRomance: false, weeklyWants:    [
-                                                                                                        {Name: "Carrot", Gifted: false}, 
-                                                                                                        {Name: "Ramen", Gifted: false}, 
-                                                                                                        {Name: "Apple Jam", Gifted: false}, 
-                                                                                                        {Name: "Dari Cloves", Gifted: false}
-                                                                                                    ]},
-                            {Name: 'Sifuu', dailyGift: false, romance: false, weeklyRomance: false, weeklyWants:   [
-                                                                                                        {Name: "Stone Brick", Gifted: false}, 
-                                                                                                        {Name: "Elder Sernuk Antlers", Gifted: false}, 
-                                                                                                        {Name: "Iron Bar", Gifted: false}, 
-                                                                                                        {Name: "Swordfin Eel", Gifted: false}
-                                                                                                    ]},
-                            {Name: 'Subira', dailyGift: false, romace: false, weeklyRomance: false, weeklyWants:   [
-                                                                                                        {Name: "Flow-Infused Wood", Gifted: false}, 
-                                                                                                        {Name: "Gold Bar", Gifted: false}, 
-                                                                                                        {Name: "Dari Cloves", Gifted: false}, 
-                                                                                                        {Name: "Firebreathing Dragonfly", Gifted: false}
-                                                                                                    ]},
-                            {Name: 'Tamala', dailyGift: false, romance: true, weeklyRomance: false, weeklyWants:  [
-                                                                                                        {Name: "Fur", Gifted: false}, 
-                                                                                                        {Name: "Garden Millipede", Gifted: false}, 
-                                                                                                        {Name: "Hairy Millipede", Gifted: false}, 
-                                                                                                        {Name: "Fairy Mantis", Gifted: false}
-                                                                                                    ]},
-                            {Name: 'Tau', dailyGift: false, romance: false, weeklyRomance: false, weeklyWants:     [
-                                                                                                        {Name: "Common Field Cricket", Gifted: false}, 
-                                                                                                        {Name: "Silver Salmon", Gifted: false}, 
-                                                                                                        {Name: "Grilled Oyster", Gifted: false}, 
-                                                                                                        {Name: "Shimmerfin", Gifted: false}
-                                                                                                    ]},
-                            {Name: 'Tish', dailyGift: false, romance: true, weeklyRomance: false, weeklyWants:    [
-                                                                                                        {Name: "Sapwood Plank", Gifted: false}, 
-                                                                                                        {Name: "Samara", Gifted: false}, 
-                                                                                                        {Name: "Petit Fives", Gifted: false}, 
-                                                                                                        {Name: "Blueberry Pie", Gifted: false}
-                                                                                                    ]},
-                            {Name: 'Zeki', dailyGift: false, romance: false, weeklyRomance: false, weeklyWants:    [
-                                                                                                        {Name: "Silvery Minnow", Gifted: false}, 
-                                                                                                        {Name: "Fish Stew", Gifted: false}, 
-                                                                                                        {Name: "Oysters Akwinduu", Gifted: false}, 
-                                                                                                        {Name: "Bouillabaisse", Gifted: false}
-                                                                                                    ]}
-]
-
-const Bugs = [         
-        [
-                { Name: "Bahari Crab", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Coast</p><p>Appears: Any Time</p><br><h1>Description</h1><p>A local crab, known to taste good with drawn butter.<br>Found along the coast of Bahari Bay. Also has a chance to be found when foraging shells in Bahari Bay.</p>"},
-                { Name: "Brushtail Dragonfly", status: "No", additionalDetails: "<p>Location: Kilima Valley, Bahari Bay</p><p>Biome: Kilima Lake, Bahari Rivers, Bahari Ponds</p><p>Appears: Any Time</p><br><h1>Description</h1><p>The tail of this Dragonfly is so bright and vibrant, that some early Majiri were said to have used it as a paintbrush.<br>Found around the lake in Kilima, as well as rivers and ponds in Bahari Bay.</p>" },
-                { Name: "Common Bark Cicada", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Fields, Bahari Forest</p><p>Appears: Morning, Day</p><br><h1>Description</h1><p>A common cicada.<br>Found attached to trees in Bahari Bay.</p>"},
-                { Name: "Common Blue Butterfly", status: "No", additionalDetails: "<p>Location: Kilima Valley, Bahari Bay</p><p>Biome: Any</p><p>Appears: Morning, Day</p><br><h1>Description</h1><p>Ever resilient. Butterflies of this type can be found in almost every climate in Palia.<br>Found all around Kilima and Bahari Bay during the morning and day.</p>"},
-                { Name: "Common Field Cricket", status: "No", additionalDetails: "<p>Location: Kilima Valley, Bahari Bay</p><p>Biome: Kilima Fields, Bahari Fields</p><p>Appears: Any Time</p><br><h1>Description</h1><p>The Common Field Cricket is a relatively common insect in Palia and can be found near Maji's Hollow, Chapaa Crossing, Whispering Banks, and around the Family Farm in Kilima Valley, as well as throughout Pulsewater Plains, amid the cliffs of Hideaway Bluffs, and outside Pavel Mines in Bahari Bay. Common Field Crickets can be caught at any time of day.</p>"},
-                { Name: "Kilima Night Moth", status: "No", additionalDetails: "<p>Location: Kilima Valley, Bahari Bay</p><p>Biome: All</p><p>Appears: Evening, Night</p><br><h1>Description</h1><p>While beautiful, the Kilima Night Moth is so prevalent on hot summer nights that some consider it to be a pest. Keep away from stored cloth.<br>Found all around Kilima and Bahari Bay during the evening and night.</p>"},
-                { Name: "Paper Lantern Bug", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Coast, Bahari Forest</p><p>Appears: Evening, Night</p><br><h1>Description</h1><p>This bug glows as brightly as a colorful feast day lantern taking flight into the Palian night sky.<br>Found along the coast and in forests of Bahari bay during the evening and night.</p>"},
-                { Name: "Spotted Stinkbug", status: "No", additionalDetails: "<p>Location: Kilima Valley, Bahari Bay</p><p>Biome: All</p><p>Appears: Any Time</p><br><h1>Description</h1><p>A bug known more for his strong musk than its striking colours.<br>Found all around Kilima and Bahari Bay.</p>"}
-        ],       
-        [
-                { Name: "Bahari Bee", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Fields, Bahari Forest</p><p>Appears: Any Time</p><br><h1>Description</h1><p>A local species of bees, known to produce delicious honey.<br>A Bahari Bee has a 5% chance to spawn after cutting down small, medium, or large Heartwood trees in Bahari Bay during any time of day.</p>"},
-                { Name: "Cerulean Cicada", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Fields, Bahari Forest</p><p>Appears: Morning, Day</p><br><h1>Description</h1><p>The call of this cicada is so soothing, it was said to be the inspiration for Ayana's fourth symphony.<br>Found attached to trees in Bahari Bay.</p>"},
-                { Name: "Duskwing Butterfly", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Dragon Shrine</p><p>Appears: Morning, Day</p><br><h1>Description</h1><p>A beautiful butterfly with wings the color of dusk. Unique to the Kilima Valley.<br>Found around the Dragon Shrine in Kilima during the morning and day.</p>"},
-                { Name: "Garden Ladybug", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Fields</p><p>Appears: Any Time</p><br><h1>Description</h1><p>A black and red ladybug, popular for controlling garden pests.<br>Has a chance to be found when foraging in the fields of Bahari Bay.</p>"},
-                { Name: "Garden Leafhopper", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Fields</p><p>Appears: Any Time</p><br><h1>Description</h1><p>The Garden Leafhopper is such a delight in Palia that many children keep them as pets. They are said to bring good luck.<br>Found in the fields and Badruu's Farm in Kilima. Also has a chance to be found when foraging in Kilima.</p>"},
-                { Name: "Garden Mantis", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Lake</p><p>Appears: Any Time</p><br><h1>Description</h1><p>Nothing to see here. Just your garden variety mantis.</p>"},
-                { Name: "Garden Millipede", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Fields</p><p>Appears: Any Time</p><br><h1>Description</h1><p>A brown and green millipede that loves feasting on garden crops.<br>Found in the fields of Kilima. Also has a chance to be found when foraging in the fields of Kilima.</p>"},
-                { Name: "Garden Snail", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Coast, Bahair Forest</p><p>Appears: Any Time</p><br><h1>Description</h1><p>A snail that has evolved over time to look like a leaf. Good for hiding from predators, bad for hiding from the Leafstalker Mantis.<br>Found along the coast and in forests of Bahari bay.</p>"},
-                { Name: "Inky Dragonfly", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Rivers, Bahari Ponds</p><p>Appears: Any Time</p><br><h1>Description</h1><p>This stealthy Dragonfly is used as a component in various black dyes and inks.<br>Found around rivers and ponds in Bahari Bay.</p>"},
-                { Name: "Lunar Fairy Moth", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Fields, Bahari Forest</p><p>Appears: Evening, Night</p><br><h1>Description</h1><p>The lunar fairy moth is said to have a very special connection to the Dragon Moon, and is said to only appear at special times during its cycle.<br>Found in the fields and forests of Bahari Bay during the evening and night.</p>"},
-                { Name: "Proudhorned Stag Beetle", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Fields</p><p>Appears: Any Time</p><br><h1>Description</h1><p>The horns of this stag beetle are said to resemble that of the Proudhorned Sernuk.<br>Has a chance to be found when mining in Kilima.</p>"},
-                { Name: "Spineshell Crab", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Coast</p><p>Appears: Any Time</p><br><h1>Description</h1><p>A crab with a very spiny shell.<br>Found along the coast of Bahari bay.</p>"}
-        ],            
-        [
-                { Name: "Azure Stonehopper", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Fields</p><p>Appears: Any Time</p><br><h1>Description</h1><p>This little blue beauty has been known to camouflage itself as a shiny blue pebble.<br>Azure Stonehoppers can be found around the fields and mountains of south Bahari Bay, including Pulsewater Plains, Hideaway Bluffs, and outside Pavel Mines at any time of day.</p>"},
-                { Name: "Bahari Glowbug", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Fields</p><p>Appears: Evening, Night</p><br><h1>Description</h1><p>A local species of bioluminescent fauna. Discovered by Professor Bonlin Amepongo.<br>Found in the fields of Bahari Bay during the evening and night.</p>"},
-                { Name: "Brighteye Butterfly", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Forest</p><p>Appears: Morning, Day</p><br><h1>Description</h1><p>Whenever you see the sun out, you can see the eyes of this intricately patterned butterfly looking around Bahari Bay.<br>Found in the forests of Bahari Bay during the morning and day.</p>"},
-                { Name: "Firebreathing Dragonfly", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Fields</p><p>Appears: Any Time</p><br><h1>Description</h1><p>This Dragonfly is notoriously difficult to catch, in part due to its fiery breath.<br>Found in the fields of Kilima.</p>"},
-                { Name: "Golden Glory Bee", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Fields</p><p>Appears: Morning, Day</p><br><h1>Description</h1><p>A rare bee, imported from the Akwinduu province. Mead made from this bee's honey is said to be heavily consumed by the queen of the province and her family.<br>Has a chance to be found when chopping trees in Kilima during the morning and day.</p>"},
-                { Name: "Gossamer Veil Moth", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Fields</p><p>Appears: Evening, Night</p><br><h1>Description</h1><p>The wings of this moth are so thin that some say you can see a peek of the Dragon Realm by looking through them.<br>Found in the fields of Kilima during the evening and night.</p>"},
-                { Name: "Hairy Millipede", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Forest</p><p>Appears: Any Time</p><br><h1>Description</h1><p>A millipede with a hair growth problem.<br>Found in the forests of Bahari Bay.</p>"},
-                { Name: "Leafstalker Mantis", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Forest</p><p>Appears: Any Time</p><br><h1>Description</h1><p>A herbivorous mantis that exclusively eats leaves.<br>Has a chance to be found when foraging in the forests of Bahari bay.</p>"},
-                { Name: "Princess Ladybug", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Lake</p><p>Appears: Any Time</p><br><h1>Description</h1><p>A very regal looking lady bug, made popular by the children's fairytale, The Princess Who Turned Into A Ladybug.<br>Found around the lake in Kilima.</p>"},
-                { Name: "Raspberry Beetle", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Daiya Family Farm</p><p>Appears: Night</p><br><h1>Description</h1><p>A beetle whose entire diet consists of raspberries.<br>Found around Badruu's Farm in Kilima.</p>"},
-                { Name: "Scintillating Centipede", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Fields, Bahari Forest, Caves</p><p>Appears: Any Time</p><br><h1>Description</h1><p>A rare centipede with a shiny hard shell.<br>Has a chance to be found when mining in Bahari Bay.</p>"},
-                { Name: "Spitfire Cicada", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Fields, Bahari Forest</p><p>Appears: Morning, Day</p><br><h1>Description</h1><p>A spirited cicada, whose bark is worse than its bite.<br>Found attached to trees in Bahari Bay.</p>"},
-                { Name: "Spotted Mantis", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Dragon Shrine</p><p>Appears: Any Time</p><br><h1>Description</h1><p>A rare mantis with a beautiful set of spots.<br>Found around the Dragon Shrine in Kilima.</p>"},
-                { Name: "Stripeshell Snail", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Coast</p><p>Appears: Evening, Night</p><br><h1>Description</h1><p>A snail with such distinctive stripes, that its shell was once used as currency among the early Majiri.<br>Found along the coast in Bahari Bay.</p>"},
-                { Name: "Vampire Crab", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Flooded Forest</p><p>Appears: Evening, Night</p><br><h1>Description</h1><p>An invasive species of crab that sucks the life out of the local population.<br>Found around the Flooded Fortress in Bahari Bay during the evening and night.</p>"}
-        ],           
-        [
-                { Name: "Ancient Amber Beetle", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Forest</p><p>Appears: Any Time</p><br><h1>Description</h1><p>One of the oldest known insect species in Palia, was said to exist in the time of humans.<br>Ancient Amber Beetles can be found in the Forests in north Bahari Bay, including the Flooded Steps, The Outskirts, Proudhorn Pass, the Thorny Thicket, and the Statue Garden at any time of day.<br>There is also a 5% chance to spawn an Ancient Amber Beetle when harvesting large or medium Sapwood trees anywhere in Bahari Bay.</p>"},
-                { Name: "Fairy Mantis", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Fields</p><p>Appears: Any Time</p><br><h1>Description</h1><p>A mantis from early Majiri legend only recently confirmed to exist by modern entomologists. Said to have magical properties.<br>Found in the fields of Bahari Bay.</p>"},
-                { Name: "Jewelwing Dragonfly", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Pavel Mines</p><p>Appears: Any Time</p><br><h1>Description</h1><p>A Dragonfly with wings as bright and colorful as any precious stone.<br>Found near a cave entrance on a cliff in Bahari Bay.</p>"},
-                { Name: "Rainbow-Tipped Butterfly", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Fields</p><p>Appears: Morning, Day</p><br><h1>Description</h1><p>Ancient legends say this rare butterfly got its beautiful coloring from dipping the tip of its wings in a Well of Flow.<br>Found in the fields of Bahari Bay during the morning and day.</p>"}
-        ]
-]
-
-const Fish = [        
-        [
-                { Name: "Albino Eel", status: "No", additionalDetails: "<p>Location: Bahari Bay, Kilima Valley</p><p>Biome: Caves</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Worm.webp'> Worm</p><br><h1>Description</h1><p>An eel with no pigmentation. Found in caves using worms.</p>"},
-                { Name: "Alligator Gar", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Lake</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>A carnivorous fish with brilliant purple scales, the Alligator Gar is named after its long snout and rows of large teeth. Found in Kilima Lake using glow worms.</p>"},
-                { Name: "Bahari Bass", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Coast</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/None.webp'> None</p><br><h1>Description</h1><p>A bountiful, staple fish known for its mild flavor. Found off the coast of Bahari Bay without bait.</p>"},
-                { Name: "Bahari Bream", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Coast</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Worm.webp'> Worm</p><br><h1>Description</h1><p>These fish are widely considered to be a staple of the local diet. Found off the coast of Bahari Bay using worms.</p>"},
-                { Name: "Bahari Pike", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Rivers</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>A pike with a deep green hue native to Bahari Bay. Found in Bahari rivers using glow worms.</p>"},
-                { Name: "Beluga Sturgeon", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Lake</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>A large sturgeon that grows throughout its entire lifetime. Found in Kilima Lake using glow worms.</p>"},
-                { Name: "Black Sea Bass", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Coast</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>Known for the black spots on their skin, these bass are gentle and non-threatening to Majiri and humans alike. Found off the coast of Bahari Bay using glow worms.</p>"},
-                { Name: "Blobfish", status: "No", additionalDetails: "<p>Location: Kilima Valley, Bahari Bay</p><p>Biome: Caves</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>A bloated, slimy fish that lives at the bottom of cavern pools. Found in caves using glow worms.</p>"},
-                { Name: "Bluefin Tuna", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Coast</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Worm.webp'> Worm</p><br><h1>Description</h1><p>A common, sleek fish adapted for swimming quickly through ocean currents. Found off the coast of Bahari Bay using worms.</p>"},
-                { Name: "Calico Koi", status: "No", additionalDetails: "<p>Location: Kilima Valley, Bahari Bay</p><p>Biome: Ponds</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/None.webp'> None</p><br><h1>Description</h1><p>A standard koi fish with a beautiful calico pattern. Found in ponds without using any bait.</p>"},
-                { Name: "Crucian Carp", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Rivers</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>A hardy carp with coppery golden scales. Found in Kilima rivers using glow worms.</p>"},
-                { Name: "Eyeless Minnow", status: "No", additionalDetails: "<p>Location: Kilima Valley, Bahari Bay</p><p>Biome: Caves</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/None.webp'> None</p><br><h1>Description</h1><p>A cave dwelling creature that has lost the need for eyes. Found in caves without bait.</p>"},
-                { Name: "Fathead Minnow", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Rivers</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>A small freshwater fish known for its oversized head. Found in Bahari rivers using glow worms.</p>"},
-                { Name: "Gillyfin", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Lake</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/None.webp'> None</p><br><h1>Description</h1><p>The most common fish found in Palia. Found in Kilima Lake without using bait.</p>"},
-                { Name: "Golden Salmon", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Rivers</p><p>Appears: Morning, Day</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Worm.webp'> Worm</p><br><h1>Description</h1><p>A freshwater salmon with golden scales. Found in Kilima Rivers using worms in the morning and day.</p>"},
-                { Name: "Kilima Catfish", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Ponds</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Worm.webp'> Worm</p><br><h1>Description</h1><p>A unique type of catfish found only in the Kilima Valley. Found in ponds using worms.</p>"},
-                { Name: "Kilima Redfin", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Rivers</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Worm.webp'> Worm</p><br><h1>Description</h1><p>An aggressive river predator that will eat literally anything. Found in Bahari Bay rivers using worms.</p>"},
-                { Name: "Largemouth Bass", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Lake</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Worm.webp'> Worm</p><br><h1>Description</h1><p>A very talkative bass that can swallow creatures over 50% their own body weight. Found in Kilima Lake using worms.</p>"},
-                { Name: "Mottled Gobi", status: "No", additionalDetails: "<p>Location: Kilima Valley, Bahari Bay</p><p>Biome: Ponds</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>An under appreciated fish that feeds on excess pond algae. Found in ponds using glow worms.</p>"},
-                { Name: "Mudminnow", status: "No", additionalDetails: "<p>Location: Kilima Valley, Bahari Bay</p><p>Biome: Ponds</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/None.webp'> None</p><br><h1>Description</h1><p>A common minnow known for its muddy looking color. Found in ponds without using any bait.</p>"},
-                { Name: "Oily Anchovy", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Rivers</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Worm.webp'> Worm</p><br><h1>Description</h1><p>A slimy fish some Majiri find strangely appetizing. Found in Bahari Bay rivers using worms.</p>"},
-                { Name: "Orange Bluegill", status: "No", additionalDetails: "<p>Location: Kilima Valley, Bahari Bay</p><p>Biome: Ponds</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Worm.webp'> Worm</p><br><h1>Description</h1><p>An omnivorous fish that prefers insects but will eat anything it can fit into its mouth. Found in ponds using worms.</p>"},
-                { Name: "Painted Perch", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Lake</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Worm.webp'> Worm</p><br><h1>Description</h1><p>A petite river fish known for its pleasing colors. Found in Kilima Lake using worms.</p>"},
-                { Name: "Platinum Chad", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Rivers</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>The brilliant platinum scales of this chad can fool fishermen into dreams of wealth. Found in Kilima rivers using glow worms.</p>"},
-                { Name: "Red-bellied Piranha", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Rivers</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Worm.webp'> Worm</p><br><h1>Description</h1><p>An aggressive river predator that will eat literally anything. Found in Kilima Rivers using worms.</p>"},
-                { Name: "Rosy Bitterling", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Rivers</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/None.webp'> None</p><br><h1>Description</h1><p>A freshwater fish with a rosy patch on its behind. Found in Bahari Rivers without bait.</p>"},
-                { Name: "Sardine", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Coast</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/None.webp'> None</p><br><h1>Description</h1><p>A small, but quick foraging fish that many Majiri enjoy eating when pickled. Found off the coast of Bahari Bay without bait.</p>"},
-                { Name: "Scarlet Koi", status: "No", additionalDetails: "<p>Location: Kilima Valley, Bahari Bay</p><p>Biome: Ponds</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>A sleek koi fish with a deep red pattern. Found in ponds using glow worms.</p>"},
-                { Name: "Silver Salmon", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Rivers</p><p>Appears: Evening, Night</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/None.webp'> None</p><br><h1>Description</h1><p>A common silvery fish. Found in Kilima Rivers without bait in evenings and nights.</p>"},
-                { Name: "Silvery Minnow", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Rivers</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/None.webp'> None</p><br><h1>Description</h1><p>A common minnow with silvery scales that sparkle in the sun. Found in Bahari Bay rivers without bait.</p>"},
-                { Name: "Smallmouth Bass", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Lake</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Worm.webp'> Worm</p><br><h1>Description</h1><p>A common bass with very little to say that is known to occasionally leap from the water. Found in Kilima Lake using worms.</p>"},
-                { Name: "Spotted Bullhead", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Rivers</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/None.webp'> None</p><br><h1>Description</h1><p>A bullhead with a spotted pattern. Found in Kilima Rivers without bait.</p>"},
-                { Name: "Stonefish", status: "No", additionalDetails: "<p>Location: Kilima Valley, Bahari Bay</p><p>Biome: Caves</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>Extremely poisonous, Stonefish naturally blend into their surroundings while waiting for prey. Found in caves using glow worms.</p>"},
-                { Name: "Striped Dace", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Rivers</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/None.webp'> None</p><br><h1>Description</h1><p>A dace with a striped pattern down its back. Found in Kilima Rivers without bait.</p>"},
-                { Name: "Yellow Perch", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Rivers</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/None.webp'> None</p><br><h1>Description</h1><p>A common perch found in rivers. Found in Bahari Bay rivers without bait.</p>"},
-                { Name: "Yellowfin Tuna", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Coast</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>A large Tuna that can swim faster than its size implies. Found off the couast of Bahari Bay using glow worms.</p>"}
-        ],       
-        [
-                { Name: "Barracuda", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Coast</p><p>Appears: Evening, Night</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/None.webp'> None</p><br><h1>Description</h1><p>A vicious saltwater predator that tastes great lightly grilled with a little butter. Found off the coast of Bahari Bay without bait in the evening and night.</p>"},
-                { Name: "Bat Ray", status: "No", additionalDetails: "<p>Location: Kilima Valley, Bahari Bay</p><p>Biome: Caves</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Worm.webp'> Worm</p><br><h1>Description</h1><p>An ominous ray that is known to leap out of the water like a flying bat. Found in caves using worms.</p>"},
-                { Name: "Cactus Moray", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Coast</p><p>Appears: Day, Night</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>An eel with sharp spines growing along its length to deter predators. Found off the coast of Bahari Bay using glow worms in the day and night.</p>"},
-                { Name: "Cantankerous Koi", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Ponds</p><p>Appears: Day, Night</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/None.webp'> None</p><br><h1>Description</h1><p>Local legend has is that the first cantankerous koi was a cranky old man transfigured by a wicked witch. Found in ponds without using any bait during the day and at night.</p>"},
-                { Name: "Channel Catfish", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Rivers</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/None.webp'> None</p><br><h1>Description</h1><p>A river catfish at home navigating fast currents. Found in Kilima Rivers without bait.</p>"},
-                { Name: "Chub", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Rivers</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>A hard-fighting fish famous for its wide ranging appetite. Found in Bahari rivers using glow worms</p>"},
-                { Name: "Crimson Fangtooth", status: "No", additionalDetails: "<p>Location: Kilima Valley, Bahari Bay</p><p>Biome: Caves</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>A small red fish with huge teeth in proportion to its own size. Found in caves using glow worms.</p>"},
-                { Name: "Cutthroat Trout", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Rivers</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Worm.webp'> Worm</p><br><h1>Description</h1><p>A river trout named after its long jaw and mouth. Found in Bahari Bay rivers using worms.</p>"},
-                { Name: "Dawnray", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Rivers</p><p>Appears: Morning</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/None.webp'> None</p><br><h1>Description</h1><p>A golden ray that hunts at dawn, its scales absorbing the sun's light. Found in Bahari Rivers without bait in the morning from 3 AM to 6 AM.</p>"},
-                { Name: "Duskray", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Rivers</p><p>Appears: Evening</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/None.webp'> None</p><br><h1>Description</h1><p>A mottled charcoal ray that hunts at dusk, blending into the riverbed. Found in Bahari Rivers without bait in the evening.</p>"},
-                { Name: "Enchanted Pupfish", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Lake</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>A cute little fish which can absorb trace amounts of Flow into its scales. Found in Kilima Lake using glow worms.</p>"},
-                { Name: "Freshwater Eel", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Rivers</p><p>Appears: Evening, Night</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Worm.webp'> Worm</p><br><h1>Description</h1><p>A long and lithe river fish often confused for a snake. Found in Kilima Village using worms in the evening and night.</p>"},
-                { Name: "Honey Loach", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Rivers</p><p>Appears: Day, Night</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>A river fish whose yellow and black stripes resemble a honeybee. Found in Bahari rivers using glow worms at day and night.</p>"},
-                { Name: "Indigo Lamprey", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Rivers</p><p>Appears: Evening, Night</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>Palians in the past would harvest the scales of this fish to make a deep indigo dye. Found in Kilima rivers using glow worms in the evening and night.</p>"},
-                { Name: "Kilima Greyling", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Rivers</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>A streamlined fish that spends the first part of its life at sea before returning to the rivers of the Kilima Valley. Found in Kilima rivers using glow worms.</p>"},
-                { Name: "Mirror Carp", status: "No", additionalDetails: "<p>Location: Kilima Valley, Bahari Bay</p><p>Biome: Ponds</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Worm.webp'> Worm</p><br><h1>Description</h1><p>A silvery carp with scales resembling mirrors. Found in ponds using worms.</p>"},
-                { Name: "Paddlefish", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Rivers</p><p>Appears: Morning, Day, Evening</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Worm.webp'> Worm</p><br><h1>Description</h1><p>A primitive fish that has survived for millions of years. Found in Bahari Bay rivers using worms.</p>"},
-                { Name: "Prism Trout", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Lake</p><p>Appears: Morning, Day</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Worm.webp'> Worm</p><br><h1>Description</h1><p>A gorgeous trout that shines like rainbows in the sunlight. Found in Kilima Lake using worms in the morning and day.</p>"},
-                { Name: "Rainbow Trout", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Rivers</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/None.webp'> None</p><br><h1>Description</h1><p>A river fish sporting a muted multicolor scale pattern. Not smart enough to be jealous of the Prism Trout. Found in Kilima Rivers without bait.</p>"},
-                { Name: "Ribbontail Ray", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Coast</p><p>Appears: Morning, Evening</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Worm.webp'> Worm</p><br><h1>Description</h1><p>Known to frequent tropical waters, the Ribbontail Ray recently made a surprising migration from the isles of Inyrevn to the west coast of the Palian Republic. Found off the coast of Bahari Bay using worms in the morning and evening.</p>"},
-                { Name: "Stalking Catfish", status: "No", additionalDetails: "<p>Location: Kilima Valley, Bahari Bay</p><p>Biome: Ponds</p><p>Appears: Evening, Night</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>A predatory blue catfish that will attack nearby vulnerable fish. Found in ponds using glow worms in the evening and night.</p>"},
-                { Name: "Stickleback", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Rivers</p><p>Appears: Morning, Day</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Worm.webp'> Worm</p><br><h1>Description</h1><p>A carnivorous fish that enjoys feasting on roe, insects, crustaceans, and the toes of fisherman. Found in Kilima Rivers using worms in the morning and day.</p>"},
-                { Name: "Striped Sturgeon", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Lake</p><p>Appears: Evening, Night</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Worm.webp'> Worm</p><br><h1>Description</h1><p>A bottom feeding lake fish with light colored stripes that allow it to camouflage in the sunlight. Found in Kilima Lake using worms in the evening and night.</p>"},
-                { Name: "Umbran Carp", status: "No", additionalDetails: "<p>Location: Kilima Valley, Bahari Bay</p><p>Biome: Caves</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/None.webp'> None</p><br><h1>Description</h1><p>A relative of the Fairy Carp, the Umbran Carp prefers the dark environments of caves and underground lake systems. Found in caves without bait.</p>"}
-        ],       
-        [
-                { Name: "Blue Marlin", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Coast</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Worm.webp'> Worm</p><br><h1>Description</h1><p>A large and resilient fish. The Blue Marlin is considered a prize by even the most experienced fishermen. Found off the coast of Bahari Bay using worms.</p>"},
-                { Name: "Blue Spotted Ray", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Coast</p><p>Appears: Morning, Day, Evening</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>A colorful stingray with bright blue spots accross its body and blue stripes along its tail. Found off the coast of Bahari Bay using glow worms in the morning, day, and evening.</p>"},
-                { Name: "Cloudfish", status: "No", additionalDetails: "<p>Location: Kilima Valley, Bahari Bay</p><p>Biome: Ponds</p><p>Appears: Morning, Day</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>A bottom feeding pondfish that can create swirling clouds of sediment for cover. Found in ponds using glow worms in the morning and day.</p>"},
-                { Name: "Energized Piranha", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Rivers</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Worm.webp'> Worm</p><br><h1>Description</h1><p>A piranha infused with magic after consuming a Grimalkin magic trinket. Found in Kilima Rivers using worms.</p>"},
-                { Name: "Fairy Carp", status: "No", additionalDetails: "<p>Location: Kilima Valley, Bahari Bay</p><p>Biome: Ponds</p><p>Appears: Morning, Evening</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Worm.webp'> Worm</p><br><h1>Description</h1><p>A shy fish with a hypnotically translucent sheen. Found in ponds using worms in the mornings and evenings.</p>"},
-                { Name: "Hypnotic Moray", status: "No", additionalDetails: "<p>Location: Kilima Valley, Bahari Bay</p><p>Biome: Caves</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>A moray eel that combines its striking pattern and hypnotic movements to lure in prey. Found in caves using glow worms.</p>"},
-                { Name: "Kenli's Carp", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Rivers</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>Kenli personally discovered this carp near Kilima Village. Bears a remarkable resemblance to a Crucian Carp... Found in Kilima rivers using glow worms.</p>"},
-                { Name: "Midnight Paddlefish", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Lake</p><p>Appears: Night</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>A rare paddlefish that feeds at night, its dark scales keeping it hidden from predators. Found in Kilima Lake using glow worms at night.</p>"},
-                { Name: "Mutated Angler", status: "No", additionalDetails: "<p>Location: Kilima Valley, Bahari Bay</p><p>Biome: Caves</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Worm.webp'> Worm</p><br><h1>Description</h1><p>A terrifying monstrosity said to be a survivor from the time of humans. Found in caves using worms.</p>"},
-                { Name: "Radiant Sunfish", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Rivers</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Worm.webp'> Worm</p><br><h1>Description</h1><p>A beautifully radiant river fish that emits a blinding ray of light many compare to looking directly into the sun. Found in Kilima Rivers using worms.</p>"},
-                { Name: "Shimmerfin", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Lake</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Worm.webp'> Worm</p><br><h1>Description</h1><p>A mysterious fish that can sometimes be in two places at once. Found in Kilima Lake using worms.</p>"},
-                { Name: "Stormray", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Rivers</p><p>Appears: Morning, Evening</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>A majestic river ray that hunts by twisting water currents and channeling lightning through its tail. Found in Bahari rivers using glow worms in the morning and evening.</p>"},
-                { Name: "Thundering Eel", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Rivers</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Worm.webp'> Worm</p><br><h1>Description</h1><p>An electrically charged eel that can shock its prey into submission. Found in Bahari Bay rivers using worms.</p>"},
-                { Name: "Willow Lamprey", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Rivers</p><p>Appears: Night</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Worm.webp'> Worm</p><br><h1>Description</h1><p>A long, thin lamprey named after its similarity to the branches of a willow tree. Found in Bahari Bay rivers using worms at night.</p>"}
-        ],          
-        [
-                { Name: "Ancient Fish", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima  Lake</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>An ancient fish preserved from the time of the lost human civilization. Found in Kilima Lake using glow worms.</p>"},
-                { Name: "Flametongue Ray", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Rivers</p><p>Appears: Morning, Day</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>A rare river ray named for the bright orange and red streaks across its back. Found in Kilima Rivers using glow worms.</p>"},
-                { Name: "Giant Goldfish", status: "No", additionalDetails: "<p>Location: Kilima Valley, Bahari Bay</p><p>Biome: Ponds</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>Against all odds, some goldfish live to become legendary in size. Found in ponds using glow worms.</p>"},
-                { Name: "Giant Kilima Stingray", status: "No", additionalDetails: "<p>Location: Kilima Valley</p><p>Biome: Kilima Lake</p><p>Appears: Morning, Day, Evening</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>The lord of Kilima Lake, this stingray's poisonous spike can also pierce the bones of its prey. Found in Kilima Lake using glow worms in the morning, day, and evening.</p>"},
-                { Name: "Long Nosed Unicorn Fish", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Coast</p><p>Appears: Day</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>The horn of the unicorn fish is valuable to any alchemist or potion master. But beware, they are rumored to have psychic abilities that can make even the most accomplished fisherman think left is right. Found off the coast of Bahari Bay using glow worms during the day.</p>"},
-                { Name: "Swordfin Eel", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Bahari Rivers</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>An apex predator eel that uses its own body to cut down prey. Found in Bahari rivers using glow worms.</p>"},
-                { Name: "Void Ray", status: "No", additionalDetails: "<p>Location: Bahari Bay</p><p>Biome: Caves</p><p>Appears: Any Time</p><p class='image-paragraph'>Bait: <img class='thumbnail' src='./src/assets/Items/Glow Worm.webp'> Glow Worm</p><br><h1>Description</h1><p>Alien and terrifying, some say the Void Ray can enter the Echo realm and feed off smaller echoes it finds. Found in caves using glow worms.</p>"}
-        ]
-]
-
-const Dishes = [
-        [
-                { Name: "Creamy Carrot Soup", status: "No", additionalDetails: "placeholder"},
-                { Name: "Fish Stew", status: "No", additionalDetails: "placeholder"},
-                { Name: "Fish Tacos", status: "No", additionalDetails: "placeholder"},
-                { Name: "Grilled Fish", status: "No", additionalDetails: "placeholder"},
-                { Name: "Grilled Meat", status: "No", additionalDetails: "placeholder"},
-                { Name: "Grilled Mushroom", status: "No", additionalDetails: "placeholder"},
-                { Name: "Grilled Oyster", status: "No", additionalDetails: "placeholder"},
-                { Name: "Hearty Vegetable Soup", status: "No", additionalDetails: "placeholder"},
-                { Name: "Hot Hounds", status: "No", additionalDetails: "placeholder"},
-                { Name: "Macaron", status: "No", additionalDetails: "placeholder"},
-                { Name: "Palian Onion Soup", status: "No", additionalDetails: "placeholder"},
-                { Name: "Sashimi", status: "No", additionalDetails: "placeholder"}
-        ],
-        [
-                { Name: "Bacon-Stuffed Mushrooms", status: "No", additionalDetails: "placeholder"},
-                { Name: "Blueberry Pie", status: "No", additionalDetails: "placeholder"},
-                { Name: "Chapaa Asada Tacos", status: "No", additionalDetails: "placeholder"},
-                { Name: "Crab Gumbo", status: "No", additionalDetails: "placeholder"},
-                { Name: "Crab Pot Pie", status: "No", additionalDetails: "placeholder"},
-                { Name: "Cream of Mushroom Soup", status: "No", additionalDetails: "placeholder"},
-                { Name: "Cream of Tomato Soup", status: "No", additionalDetails: "placeholder"},
-                { Name: "Fisherman's Brew", status: "No", additionalDetails: "placeholder"},
-                { Name: "Fried Catfish Dinner", status: "No", additionalDetails: "placeholder"},
-                { Name: "Loaded Potato Soup", status: "No", additionalDetails: "placeholder"},
-                { Name: "Ramen", status: "No", additionalDetails: "placeholder"},
-                { Name: "Sernuk Noodle Stew", status: "No", additionalDetails: "placeholder"},
-                { Name: "Shepp's Pie", status: "No", additionalDetails: "placeholder"},
-                { Name: "Steak Dinner", status: "No", additionalDetails: "placeholder"},
-                { Name: "Trout Dinner", status: "No", additionalDetails: "placeholder"}
-        ],
-        [
-                { Name: "Akwinduu Chapaa", status: "No", additionalDetails: "placeholder"},
-                { Name: "Apple Pie", status: "No", additionalDetails: "placeholder"},
-                { Name: "Bouillabaisse", status: "No", additionalDetails: "placeholder"},
-                { Name: "Chapaa Onigiri", status: "No", additionalDetails: "placeholder"},
-                { Name: "Chili Oil Dumplings", status: "No", additionalDetails: "placeholder"},
-                { Name: "Meaty Stir Fry", status: "No", additionalDetails: "placeholder"},
-                { Name: "Phoenixfire Relleno", status: "No", additionalDetails: "placeholder"},
-                { Name: "Spicy Stir Fry", status: "No", additionalDetails: "placeholder"},
-                { Name: "Sushi", status: "No", additionalDetails: "placeholder"},
-                { Name: "Veggie Fried Rice", status: "No", additionalDetails: "placeholder"}
-        ],
-        [
-                { Name: "Celebration Cake", status: "No", additionalDetails: "placeholder"},
-                { Name: "Chapaa Masala", status: "No", additionalDetails: "placeholder"},
-                { Name: "Poke Bowl", status: "No", additionalDetails: "placeholder"},
-                { Name: "Spicy Rice Cakes", status: "No", additionalDetails: "placeholder"}
-        ],
-                { Name: "Petit Fives", status: "No", additionalDetails: "placeholder" },
-                { Name: "Oysters Akwinduu", status: "No", additionalDetails: "placeholder" },
-                { Name: "Muujin Bahari", status: "No", additionalDetails: "placeholder" },
-                { Name: "Congee", status: "No", additionalDetails: "placeholder" }
-]
-
-// https://palia.wiki.gg/wiki/Furniture/Sets_Catalogue
-        const Furniture = [
-                                { Name: 'Bellflower', pieces:    [ 
-                                                                        { Name: "Alarm Clock", status: "No", additionalDetails: "placeholder" },
-                                                                        {Name: "Armchair", status: "No", additionalDetails: "placeholder" },
-                                                                        {Name: "Banner", status: "No", additionalDetails: "placeholder" },
-                                                                        {Name: "Bathtub", status: "No", additionalDetails: "placeholder" },
-                                                                        {Name: "Bed", status: "No", additionalDetails: "placeholder" },
-                                                                        {Name: "Bookshelf", status: "No", additionalDetails: "placeholder" },
-                                                                        {Name: "Chandelier", status: "No", additionalDetails: "placeholder" },
-                                                                        {Name: "Coffee Table", status: "No", additionalDetails: "placeholder" },
-                                                                        {Name: "Couch", status: "No", additionalDetails: "placeholder" },
-                                                                        {Name: "Curtains", status: "No", additionalDetails: "placeholder" },
-                                                                        {Name: "Dining Chair", status: "No", additionalDetails: "placeholder" },
-                                                                        {Name: "Dining Table", status: "No", additionalDetails: "placeholder" },
-                                                                        {Name: "Dresser", status: "No", additionalDetails: "placeholder" },
-                                                                        {Name: "Flower Planter", status: "No", additionalDetails: "placeholder" },
-                                                                        {Name: "Grand Fountain", status: "No", additionalDetails: "placeholder" },
-                                                                        {Name: "Mirror", status: "No", additionalDetails: "placeholder" },
-                                                                        {Name: "Nightstand", status: "No", additionalDetails: "placeholder" },
-                                                                        {Name: "Pantry", status: "No", additionalDetails: "placeholder" },
-                                                                        {Name: "Picture Frame", status: "No", additionalDetails: "placeholder" },
-                                                                        {Name: "Reed Planter", status: "No", additionalDetails: "placeholder" },
-                                                                        {Name: "Regal Fountain", status: "No", additionalDetails: "placeholder" },
-                                                                        {Name: "Rug", status: "No", additionalDetails: "placeholder" },
-                                                                        {Name: "Runner", status: "No", additionalDetails: "placeholder" },
-                                                                        {Name: "Shrub Planter", status: "No", additionalDetails: "placeholder" },
-                                                                        {Name: "Sink", status: "No", additionalDetails: "placeholder" },
-                                                                        {Name: "Small Rug", status: "No", additionalDetails: "placeholder" },
-                                                                        {Name: "Standing Lamp", status: "No", additionalDetails: "placeholder" },
-                                                                        {Name: "Table Lamp", status: "No", additionalDetails: "placeholder" },
-                                                                        {Name: "Toilet", status: "No", additionalDetails: "placeholder" },
-                                                                        {Name: "Wall Cabinet", status: "No", additionalDetails: "placeholder" },
-                                                                        {Name: "Wall Shelf", status: "No", additionalDetails: "placeholder" },
-                                                                        {Name: "Wardrobe", status: "No", additionalDetails: "placeholder" }
-                                                                ]
-                                },
-                                { Name: 'Capital Chic', pieces:       [ 
-                                                                        { Name: "Armchair", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Bathtub", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Bed", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Cabinet", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Chandelier", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Coffee Table", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Couch", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Dining Chair", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Dining Table", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Doormat", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Fern Planter", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Field Frame", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Kitchenette", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Large Shelf", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Mirror", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Nightstand", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Pantry", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Planter", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Mirror", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Rug", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Shelves", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Sink", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Small Shelf", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Stylish Rug", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Table Lamp", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Wall Cabinet", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Wall Decor", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Wall Shelf", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Wardrobe", status: "No", additionalDetails: "placeholder" }
-                                                                ]
-                                },
-                                { Name: 'Chapaa Hut', pieces:       [ 
-                                                                        { Name: "Bahari Board", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Chapaa Mug", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Embra Plush", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Enkii Pillow", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Enkii Plush", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Fun Guy Plush", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Inki Pillow", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Inki Plush", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Maji Plush", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Mini Maji Plush", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Naanshi Pillow", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Naanshi Plush", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Neena Pillow", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Neena Plush", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Nergaal Pillow", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Nergaal Plush", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Pillow", status: "No", additionalDetails: "placeholder" }
-                                                                ]
-                                },
-                                { Name: 'Dragontide', pieces:       [ 
-                                                                        { Name: "Abstract Rug", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Armchair", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Bathtub", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Bed", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Bookshelf", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Chandelier", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "Coffee Table", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" }
-                                                                ]
-                                },
-                                { Name: 'Emberborn', pieces:       [ 
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" }
-                                                                ]
-                                },
-                                { Name: 'Garden Variety', pieces:       [ 
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" }
-                                                                ]
-                                },
-                                { Name: 'Homestead', pieces:       [
-                                                                        { Name: "Alarm Clock", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" }
-                                                                ]
-                                },
-                                { Name: 'Industrial', pieces:       [ 
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" }
-                                                                ]
-                                },
-                                { Name: 'Kilima', pieces:       [ 
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" }
-                                                                ]
-                                },
-                                { Name: 'Kilima Rummage Piles', pieces:       [ 
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" } 
-                                                                ]
-                                },
-                                { Name: 'Log Cabin', pieces:       [ 
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" } 
-                                                                ]
-                                },
-                                { Name: 'Maji Market', pieces:       [ 
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" } 
-                                                                ]
-                                },
-                                { Name: 'Makeshift', pieces:       [ 
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" } 
-                                                                ]
-                                },
-                                { Name: 'Moonstruck', pieces:       [ 
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" } 
-                                                                ]
-                                },
-                                { Name: 'Ranch House', pieces:       [ 
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" } 
-                                                                ]
-                                },
-                                { Name: 'Ravenwood', pieces:       [ 
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" } 
-                                                                ]
-                                },
-                                { Name: 'Spooky', pieces:       [ 
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" } 
-                                                                ]
-                                },
-                                { Name: 'Valley Sunrise', pieces:       [ 
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" },
-                                                                        { Name: "placeholder", status: "No", additionalDetails: "placeholder" } 
-                                                                ]
-                                }
-                        ]
-
-export { Villagers, Bugs, Fish, Dishes, Furniture}
